[engine] Check that types passed to TypeConstraint inits are in fact types
I was adding some type constraints somewhere and didn't splat types. This failed to work because a list is not a type.

This patch adds a validation that raises a TypeError if the types are not types

Testing Done:
Wrote a failing test and made it pass. CI away in the PR.

Bugs closed: 3841

Reviewed at https://rbcommons.com/s/twitter/r/4209/

closes #3841