Add description to type constraints
In cases when a type constraints types are multitudinous, it's nice to have repr / str display a human friendly name instead of the long list of types.

It's not used as of this patch, but it's a distinct optional dependency of a larger patch I'm working on. In that patch, I'm adding type constraints with the type contents of the BUILD file symbol tables. These are long enough and not obviously about parsing that I felt like having this is worth while.

Testing Done:
Added tests covering Exactly's repr and strs.

Bugs closed: 3865

Reviewed at https://rbcommons.com/s/twitter/r/4233/

closes #3865