Inject an automatic dep on junit for all junit_tests targets.
It's silly to require an explicit dep.

This required creating a new "JUnit" subsystem.  While there,
I moved the junit runner tool to that subsystem, so the runner
and the library are in the same place.

It's still clunkier than it needs to be to inject deps onto targets,
but that's a problem for another time.

This change removes the explicit junit dep from one of our examples,
to prove that this all works.  A subsequent change will remove any
other unnecessary explicit deps on junit.

Testing Done:
CI passes: https://travis-ci.org/pantsbuild/pants/builds/159432189

Reviewed at https://rbcommons.com/s/twitter/r/4228/