Get rid of the 'codegen' target label and associated checks.
Instead, use type checks on a new CodegenLibraryMixin, created
to tag codegen libraries.

Note that this mixin lives in src/python/pants/build_graph,
as this is a good central point for it.  Putting it in, say,
the codegen backend, would require other backends to depend on
codegen, just to import the mixin. Dependencies between backends
is something we'd like to avoid as much as possible.

This is part of an effort to get rid of the long-deprecated
target labelling system, and its associated is_* methods.