Allow attaching a `prep_command` to several goals.
Previously only a single goal could be specified leading to copy/paste
boilerplate targets and most likely a target aggregator when the
`prep_command` needed to be run in multiple goals with no common
ancestor. This change adds the `goals` kwarg to `prep_command` and
deprecates `goal` to allow reduction of boilerplate.