Remove the long-deprecated Target.is_codegen().
Replace all calls with calls to Tqrget.is_synthetic().

In pretty much all cases where we need to check if a target is a codegen
target, what we actually care about is whether that target was synthesized
in order to hold generated code.  If didn't use target synthesis as an
implementation detail, we probably wouldn't have needed those is_codegen
checks to begin with.