Prepare the 1.2.0 release
This prepares the 1.2.0 release, which is a noop relative to 1.2.0rc4 from Friday.

There was also a bit of ambiguity in the release documentation about whether the `1.X.Y` stable notes should contain the dev releases from the series as well. From a stable-release-consumer's perspective, the goal will generally be to "see all changes since the previous stable release", and including the unstable releases from the series on the page helps with that.

So, in order to make things a bit easier for stable consumers, this removes that ambiguity in favor of including the `dev` releases.

- Declare the `1.2.0` release.
- Clarify that when creating notes for a new stable branch, you should bootstrap them with the relevant `dev` releases.
- Fixup headers for `1.2.x.rst`, `1.1.x.rst`, and `1.0.x.rst` to align with the above advice.
- Include the `1.2.0devX` notes in the `1.2.x` notes page (the `1.1.xdevX` releases were already included on `1.1.x`).

Testing Done:
https://travis-ci.org/pantsbuild/pants/builds/172105942

Bugs closed: 4008

Reviewed at https://rbcommons.com/s/twitter/r/4341/