Re-work native engine version.
This switches the `Native` subsystem to load the proper native engine
library version from a resource file that is, in-turn, auto-managed by
the pants script itself ensuring that pants, both in-repo and as a
released package (pex or sdist or wheel), will always ask for the exact
native engine library version it was built and tested with.

A follow-up change will ensure we release a native engine library to
bintray for all packaged releases, fulfilling the released pants package
portion of the contract.