Patch to make scala tests work
Change to make scalatest test work using the JUnitRunner builtin to scala test.  Basically I check if the test extends org.scalatest.Suite, if it does I include it in the specs.  Later on, we do the same thing when converting the test class to a junit Runner.  We check to see if it is a Suite, and if so then fake like we included the @RunWith annotation.

A few notes:
  1) I didn't support scala test "methods"
  2) I used reflection instead of modifying junit.py and friends to make scalatest be a non-shaded jar (in fact we don't need to include it at all)
  3) I made sure I used the test class's class loader to load Suite and JUnitRunner
  4) I bumped the runner_jar version to 1.0.16, I hope this is correct.

I would love to see this as a 1.2.1 release.

NOTE:  This is a recreate of #4344

Bugs closed: 4013

Reviewed at https://rbcommons.com/s/twitter/r/4361/