Move SimpleCodegenTask into the pants core.
This allows us to do away with the 'codegen' backend,
in favor of a thrift backend, proto backend, etc.
The idea of a single global codegen backend makes no sense,
as they share almost no code.

The one part they do share, SimpleCodegenTask, is small,
self-contained, and represents a generic enough concept
that it can comfortably live in the core.

This change also modifies SimpleCodegenTask's test to not
rely on the codegen or JVM backends.