Restore propagation of thrown exceptions between rust and python (#4083)
### Problem

In the first implementation of the rust engine, I punted on passing thrown exception values back and forth between rust and python, to avoid dealing with string manipulation. A fair number of tests were disabled in order to account for that.

### Solution

Rust `Throw` states now hold a `Value` representing a python exception, and a new extern was created to allow rust to instantiate an Exception for errors occurring in scheduling.

### Result

Throw values always contain python objects, #4025 is fixed, and all tests that were disabled for #4025 are now reenabled.