Allow the buildroot to be a source root.
This is of limited use, because it would require all
languages in your repo to be rooted at the buildroot.
But I did encounter one such case (involving .proto files,
which Google now recommends be imported relative to the
buildroot), so might as well support it.

TODO: This brings to the fore the fact that source root
      computation is not dependent on the target type.
      If it were (e.g., java_library would only match build roots
      that are registered against 'java') then you could have
      .protos rooted at the buildroot but java in src/java etc.
      See https://github.com/pantsbuild/pants/issues/4093.