[engine] Minor perf: use _kwargs in Struct._key
Problem
-------
Struct._key is called ~460k times running list in the pants repo. The @property machinery invoked for self.address ends up taking ~3-5% of time.

Solution
--------
Inline the _kwargs.get call for address in this specific case.

Another more invasive solution would be to set __dict__ to _kwargs, but that would have other effects that we might not want.