Reduce hashing during v2 transitive graph walks (#4109)
### Problem

Transitive graph walks in the v2 engine involve doing lots of deduping merges of collections of `HydratedTarget` objects. Pre-change, this was taking up about 70% of the total runtime of `./pants --enable-v2-engine list $target` for a highly connected `$target`.

### Solution

OrderedSet is significantly slower for individual dedupe operations (particularly when it is converted back into a tuple afterward), so we switch to deduping a generator using a throwaway set and collecting it into a tuple.

Additionally, because `HydratedTarget` objects are guaranteed to have an Address, we implement equality/hash checks using an Address lifted from the inner structs.

### Result

The runtime of `./pants --enable-v2-engine list $target` is improved by approximately 2x.