Bump pex and setuptools to latest. (#4111)
Problem

Pants uses an old version of setuptools (as of last week, 5.x and currently at 20.x). We'd like to modernize our consumption of this lib as composed with pex to stay current with improvements, bugfixes, etc around working with python requirements and resolution.

Solution

This change consumes a new version of pex which extends the higher end of the setuptools requirement range to permit usage of the latest version of setuptools (30.0.0) and also bumps the setuptools dep to the latest.

Result

Tested this end-to-end in both pants' and Twitter's monorepo with clean results.