Execute traces for all non-Return values (#4118)
### Problem

Currently, `Noop`s are not returned across the native boundary, and thus aren't traced during a failure. Noops can occur for a few reasons, but the most user-facing of those reasons are build graph cycles.

### Solution

Rather than executing traces only when we see a Throw, execute it in all non-Return cases, which handles Noops and other unexpected conditions.