[engine] model snapshots in validation, make root rules a dict instead of a set (#4125)
### Problem
I'm working on porting my patch for #3580 to the native engine, (#3960). First, I'm updating the graph construction code with the changes there. It hadn't covered Snapshot processes, and it modeled root rules as a set rather than a graph, which caused problems.

### Solution

Include snapshot process' behavior in the rule graph construction used by the validator. Update tests to expect the new visualization format.

### List of changes
* rename NodeBuilder to RuleIndex
* rename RootRule to RootRuleGraphEntry
* support SnapshotProcess in validator
* Name snapshot process funcs so they are printable
* Include subject and product types in intrinsic rule repr
* Include graph generation methods that can construct new graphs from an existing one--these are currently unused
* Raise an error if a declared rule is not reachable from the declared root types.
* Unify naming for input selector for projection selectors
* If field of a SelectDependencies is the default value, don't include it in the repr
* Include a map of root rule to the rules that would fulfill it.