Redirect bootstrapping calls in pants binary to stderr (#4131)
### Problem
CI fails intermittently in the pre-commit checks while attempting to run `isort.sh`. My hypothesis is that it is because the bootstrap and activation parts of the pants script are echoing things to stdout. `isort.sh` runs `./pants changed` and tees the results into a file, assuming that only target specs will be added to the file. If bootstrap were to write to stdout, it would cause problems when the specs were passed to the fmt.isort call.

### Solution
This patch redirects the activation and bootstrap calls' stdout to stderr which ensures they will not interfere even if they have writes to stdout.