Raise an Exception if relative_symlink tries to unlink a concrete file
As said in the comment, this has been a long-term uncaught exception,
surfacing when os.unlink was called on a directory. This resulted
in non-zero exit for Pants, with Operation not Permitted raising
from the OS.

Should now give a decent msg, but the outcome of Pants in this
situation will continue to be an execution-halting exception.

This behavior has diverged between relative and absolute symlinks
methods. the absolute_symlink checks for dirs, and silently rms
them. That behavior should probably be reconciled, one way or another,
one of these days.

Once a determination is made, it would likely inform the final determination
on whether safe_mkdir should be allowed to continue to be symlink agnostic.