Only copy over previous results after checking for cache hits
This was probably really clogging the IO, since every jvm
target was:

* copying the entire previous_results_dir over
* doing it a second time
* checking for cached artifact
* if found, parsing the analysis and pruning unneeded results.

Now it creates the current VT results_dir, and then checks for the
cache. This includes the double check, so the last thing that is
done before yielding the VT as invalid is to copy any
available previous results.

My read on this is that it will save a good chunk of unnecessary system
calls.