Python lint error
WHoops. I have gotten so used to disabling the style checks while
developing!

Otherwise appears to be good. There is still not enough test coverage
around invalidation in general, and incremental tasks specifically.
But I have added a lot more then there used to be.

I have a logic refactor as a followup, and I will try and
continue to add coverage as a part of that.