Add some extra tests and error message to vt.ensure_legal
At the moment this throws an error in these circumstances.
This is probably a win from a UX perspective, since we will
finally get some caught exceptions if someone starts
deleting dirs under pants.d. But perhaps you would prefer
to silently restore the working state for now, so let me
know if that is the case.

Also removes no longer needed method and retitles the
InvalidResultsDir() exception to avoid overloading "invalid" any
more than it already is.