Cleanup unused cffi handles to free memory (#4135)
### Problem

Previously, the set of CFFI handles in the `ExternContext` would grow unboundedly as they were created, amounting to all `Values` that were used (and which later became unused) being leaked.

### Solution

To allow CFFI handles to be collected, `Clone` and `Copy` were removed from the rust `Value` type, which put restrictions on who could create/copy Values. `Drop` was then implemented for `Value` such that a Value going out of scope caused its internal handle value to be enqueued to later have its underlying handle dropped on the python side (potentially allowing it to be garbage collected).

It's unavoidable that multiple CFFI handles will exist per underlying object, because we don't want to hash all Values to dedupe them (we only hash objects that we are forced to due to lifting to a Key). Because of this, doing reference counting on the rust side (although easy enough) would have been confusing due to both reference counting and multiple copies existing.

### Result

Memory previously leaked in `ExternContext._handles` should be periodically freed instead. Unfortunately, the additional overhead of tracking usage causes a slight (5-10%) slowdown.