[docs] Update the cache section on the Task developer page.
The existing example had a bug, in that it was only checking the cache
for a VTS of the invalid_vts.

I also simplified the example to only the pertinent parts. It is possibly
marginally not optimal, in that it doesn't update the cache with all
pairs at once, but once per vt. But we do this within jvm_compile,
so it isn't like it doesn't happen. And I think that this is much
clearer about what needs to happen, since it doesn't offload the
pair construction to an unimplemented match_up function.

I believe the examples are a bit clearer and the section is about
25 lines shorter as a bonus.