Add a predicate the the example target gathering
By request, and I agree. Only core tasks should not filter by
Target type. Anyone reading these docs should filter, this makes
it clear.

I prefer to use the predicate=fn personally, but my goal with the
sample code was for it to be complete and run if pasted in a
task. Even something like self._is_your_target is hand wavy if
you are a first time task developer.