Fork to post_stat (#4170)
Problem:

Earlier thread to post_stat may leave the pants process hanging from exiting until timeout. Although demonizing then joining the thread may help, post_stat procedure is already at the very end of the pants run, so there is not much time to save if we eventually want to wait for it to finish.

Solution:
Fork a process to do the post.