Reduce compilation invalidation scope of targets with strict_deps=True (#4143)
### Problem

Related issue: #3060 

Now, with strict_deps turned on, only direct dependencies will be used to compile jvm targets. But it isn't taken into account when generating fingerprints for compile. We can reduce the invalidation scope as well.

An example will be following:
Let's assume we have 3 targets A, B, and C. A depends on B; B depends on C, and A does not directly depends on C (C is a transitive dependency of A). In current implementation, invalidation hash of A takes hashes of A, B and C into account. But with strict_deps turned on, changing C does not have to invalidate A, since A does not depend on C for compilation.

### Solution

I add a new "direct(self, target)" method in FingerpringStrategy abstract base class. This method returns "False" by default, derived class of FingerpringStrategy can override this method. In the case of this review, ResolvedJarAwareTaskIdentityFingerprintStrategy overrides this method so that when a target's strict_deps is defined then return its value, otherwise, platform's strict_deps property will be used. The target.transitive_invalidation_hash() method will use direct() method to determine the scope of invalidation hash.

 This review also adds a unit test which verifies direct invalidation logic works as expected and several integration tests.