Problem
Scrooge codegen does not do fingerprinting for enough facets and is
missing support for include_path and default-java-namespace.

Solution

Add support for include-path and default-java-namespace.

Fix the fingerprinting to include namespace_map, include_path and
default_java_namespace.

Result

Scrooge support is now good enough for to be used for development
of Scrooge itself.