Add __init__.py for tests/python directories (#4193)
We have most of our test directories containing an `__init__.py`, a few probably we forgot. The consequence of that is `pytest` expected either all w/ `__init__` or all without, when mixed together the behavior is assumed by looking at the first test file, whether the directory it's in contains `__init__.py`, the assumed behavior may not work for the other, this causes flakiness when input files ordering changes.

As an example, turning on v2 changes the list ordering and sometimes happens to put `tests/python/pants_test/invalidation/test_strict_deps_invalidation_integration.py` as the first test file to collect, `tests/python/pants_test/invalidation/__init__.py` does not exist. For the rest tests we see "ImportError: `No module named test_export_classpath_integration` [1]

The alternative is to remove `__init__.py` [2] entirely for all tests/python. I tried that, the issue is our tests directories are not organized in `1:1:1` style, those non-test packages under the same directories won't be recognized properly.

So we either adding `__init__.py` which is what this review does or do the refactoring of test libraries.

[1] https://travis-ci.org/peiyuwang/pants/jobs/193144087#L3859
[2] http://doc.pytest.org/en/latest/goodpractices.html "avoid “__init__.py” files in your test directories"