A task to generate Python code from ANTLR3 grammars.
This is intended to supercede the existing AntlrBuilder
in the old Python pipeline. It does not add new features
and does not share any code with the Java ANTLR task.
That is left as future work.

Also rearranges the old ANTLR testprojects, removes cruft
and adds various tweaks and TODOs to the Java ANTLR task.