Qualify the build invalidator root.
Adds the name of the fprint strategy, and the value of the invalidate_dependents
setting, to the build invalidator root path.

Previously a task had a single root for all fprint strategies and transitivity
settings it might use.  Mostly this isn't a problem, since a task typically uses
only one such combination.

However the BootstrapJvmTool uses two: the ShadedToolFingerprintStrategy for
shaded tools and the IvyResolveFingerprintStrategy for unshaded ones.
If you happen to use the same jar spec as both a shaded and unshaded tool
then the two different fingerprints will thrash on the same key file, causing
unnecessary and confusing invalidation.

This isn't hypothetical: We have a jar that acts both as a custom javac compiler jar
and also contains other tooling. Pants doesn't shade the compiler so plugins can access
its internals, but we do shade the jar when used for other tooling, for the usual reasons.

This also removes a wrinkle where we were creating a BuildInvalidator directly
just to call force_invalidate_all() on it. That is no longer a good abstraction, since
we can't at that point know what the qualifying subdirs are. Fortunately, however,
we don't actually need that functionality.  It was only used in one place in a test
base class, and we know that each test now runs in a clean workdir anyway.

This change also gets rid of the helper function to create a cache manager, since
nothing outside the class was actually using it.  So this change has the nice
side-effect of further encapsulating both CacheManager and BuildInvalidator.