Derive object id used in the native context from object's content (#4233)
### Problem

In the native context, object uses an integer identifier which is derived by checking hash existence. In case of user overriding hash function, this can be buggy because object change may not be reflected and may end up retrieving stale object.

#4199 illustrates the problem in more detail. Note `Struct` as of today is not used as a subject, neither does any of the subject objects overrides their hash function, but this is still a potential risk.

### Solution

Switch to use the pickled content digest, still keep the integer as the external facing identifier to live in the native context, for the purpose of hiding the implementation detail. As in this review, it's one to one: digest to integer id.

### Result

Performance impact estimate: negative `~40%`