[engine] Begin port of engine rule graph validation to Rust (#4227)
This is the first step towards adapting the static noop culling I worked on last year to the Rust code base.

This first patch is working towards replacing the graph validator written in python with one in rust. At this point it covers quite a lot of behavior, but it isn't working quite correctly yet.

This patch includes porting the majority of the rule graph construction. Including
- replacing the root selector fn dict with a set of subjects and a method. Since all of the subjects now use the same selector, it didn't make sense to keep.
- running the rust validator, but ignoring the results.
- adding copyright headers.

rule_graph.rs is where the meat of the port is. There are still a number of unported sections. I'll be removing the commented python and replacing it with rust in subsequent pull requests

The remaining portions are
- adding a enum on tasks for their task type
- removing unfulfillable entries from the graph
- handling SelectDependencies / SelectProjection selectors
- handling snapshots
- propagating the results back to the python side