[engine] fix address node creation in v2 build graph; fix filedeps (#4235)
### Problem

$ ./pants --enable-v2-engine filedeps abc::
...
Exception message: 'Address' object has no attribute 'build_file'
In the v1 engine an Address is resolved to a BuildFileAddress, in the v2 engine thats not the case. Hence, invocation of goals such as filedeps causes an exception.

### Solution

Make BuildFileAddress a first class citizen. Address/BuildFileAddress can really be the same thing. To reduce the complexity of this change, split the refactor to another rb.

### Result

Remove BuildFile usage in v2 path.