[engine] Porting validation to Rust pt ii (#4243)
This patch ports the rest of the static graph generation algorithm to rust.

It hooks up printing errors, so that if an invalid rule is added it will be printed twice. I tested it manually. The next step will be to integrate with the existing validation tests and remove the python implementation.

#### Highlights:
* **pass field type info for SelectDependencies** added a new struct type for passing lists of TypeIds to support the field_types element.
* **cull unfulfillable rules transitively** In rule_graph.rs, `_remove_unfulfillable_rules_and_dependents` is ported, which transitively removes unfulfillable rules from the graph.