[engine] Use Value in invoke_runnable as the function instead of Function (#4258)
### Problem

Sometimes we want to be able to invoke Python functions from Rust that are not pickleable. Right now, we use the `Function` struct. It requires a `Key` to have been constructed for the function in order to use it. Building keys requires pickling, so unpickleable functions cannot be called from Rust.

### Solution

Use Value instead of Function for runnable calls. Value doesn't have to require a pickle, so unpickleable functions can be represented with it.

The one funny bit here is `Key::new_with_anon_type_id`. `val_for` doesn't check the type id of the passed key, so this works. It could be cleaner though.