[engine] Make Graph.get generic and make Externs static (#4261)
### Problem

Currently, any `Node` running in the `Graph` results in a `Value`, but this limits what can be stored in the Graph to things that can be represented as python objects. Moving forward (particularly for #4172), we want to be able to store rust types directly in the graph, to avoid roundtrips to python in many cases.

Additionally: the `Externs` class (despite representing 100% static resources on the python side) was not stored statically on the rust side. This meant that it had to be passed into a bunch of contexts explicitly, which (in particular) made it impossible to directly implement `Clone`/`Debug` for `Value`. 

### Solution

Rename `Node` to `NodeKey` (since it is used as both the unique identifier and inputs of a Node), rename `Step` to `Node`, since it encompasses the relevant requirements of a `Node` better than the struct does, and Make `Graph.get` generic on the type of `Node`. A `Node` implementation can be converted `Into` a `NodeKey`, and its result can be consumed out of `NodeResult` (which currently only contains `Value`, but will contain rust types in future).

Also, `Externs` were made static, and all utility methods therein are now accessed statically. A few traits (`Clone`/`Debug`) can now be utilized directly for `Value` and its containing types.

### Result

Easier usage of Externs, and important preparation for #4172.