Make options fingerprinting very difficult to disable (#4262)
### Problem

Options fingerprinting as provided by `TaskIdentityFingerprintStrategy` has proven to be extremely useful. But since passing a `fingerprint_strategy` to `Task.invalidated` acts as an _override_ of the default strategy (rather than being additive), it is currently easy to accidentally disable options fingerprinting (as happened in #4177 when a `Task` began passing a strategy).

When options fingerprinting was initially introduced, we were less certain about its efficacy and risks, which likely led to the decision to make it very easy to disable fingerprinting Task options. But in hindsight, options fingerprinting is the right choice in all cases we can think of.

### Solution

Make options fingerprinting very difficult to disable by pushing it down to the `CacheKeyGenerator`, which is initialized once per `Task`. Deprecate `TaskIdentityFingerprintStrategy`, as it is now redundant.

### Result

Options that are marked `fingerprint=True` will always be included in a `Task`'s fingerprint.